+++ src/Context/Options/WordPressOptionsContext.php
@@ -1,0 +1,26 @@
+<?php
+namespace StephenHarris\WordPressBehatExtension\Context;
+
+class WordPressOptionsContext implements Context
+{
+
+    /**
+     * @Given I set :option option to :value
+     */
+    public function iSetOptionTo($option, $value)
+    {
+        update_option($option, $value);
+    }
+
+    /**
+     * @Then I the :option option should be set to :value
+     */
+    public function theOptionShouldBeSetTo($option, $value)
+    {
+        $actual = get_option($option);
+
+        if ($actual != $value) {
+            throw new \Exception(sprintf('Expected option to be set to "%s". Found "%s"', $value, $actual));
+        }
+    }
+}