+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project is currently at an alpha stage. It will continue on the `0.*.*` bra
 - Extracted `WordPressPostContext` from `WordPressContext` and extracted helper methods into `WordPressPostRawContext` trait
 - Extracted `WordPressTermContext` from `WordPressContext` and extracted helper methods into `WordPressTermTrait` trait
 - Extracted `WordPressUserContext` from `WordPressContext` and extracted helper methods into `WordPressUserTrait` trait
+- Extracted `Plugins\WordPressPluginsContext` from `WordPressContext`
+- Extracted `Options\WordPressOptionsContext` from `WordPressContext`
+- Extracted `WordPressUserContext` from `WordPressContext` and extracted helper methods into `WordPressUserTrait` trait
 - Extracted `WordPressEditPostContext` from `WordPressAdminContext`
 - Extracted a `AdminPage` page object from `WordPressAdminContext` (see http://behat-page-object-extension.readthedocs.io/en/latest/index.html)
 - Replaced `When I hover over the row containing :value in the :column_text column of :table_selector` to `When I hover over the row containing :value in the :column_text` (`:table_selector` removed, and is internally set to `.wp-list-table`